feat: implement analytics issues #246 #247 #248 #249
- Add useRefresh hook and RefreshIndicator component (#246)
- Add ComparisonTable component, use in comparison page (#247)
- Add EmailPreview component and email-template lib (#248)
- Fix chart lazy loading: remove duplicate imports, fix JSX nesting (#249)
