+++ analytics/lib/email-template.tsx
@@ -1,0 +1,7 @@
+export const EMAIL_SUMMARY_CARDS = [
+  { label: 'Total gists', value: '18,420' },
+  { label: 'Active users', value: '6,320' },
+  { label: 'Engagement rate', value: '24.8%' },
+] as const;
+
+export const DASHBOARD_URL = 'https://github.com/PinSpace-Org/GistPin';
